feat: Unify artist bubble avatars across dashboard, sidebar, and browse page
Extract shared gradient+initials helpers to artist-avatar.ts. Right sidebar
Top Artists, dashboard Recommended/Recently Added, and browse page all now
use the same circular gradient bubble style.
